feat(graph): produce bundle-bound sources for single repos
