+++ docs/tasks/board.md
@@ -48,3 +48,4 @@
 | TASK-GRAPH-CURRENT-STATE-AUDIT-001 | Graph Current-State Audit | done | `docs/diagnostics/graph-current-state-audit.md`, `merger/lenskit/tests/test_graph_current_state_audit.py`, `docs/proofs/graph-current-state-audit-proof.md` | Inventarisiert Graph-, Entrypoint- und Graph-Index-Contracts, Producer, CLI, bedingte Bundle-Integration, Stale-Prüfung und opt-in Retrieval-Nutzung. Belegt als Priorität, dass `stale_or_mismatched` derzeit trotz erkanntem Hash-Mismatch das Ranking verändert. Keine Runtime-Mutation in diesem Audit-Slice. |
 | TASK-GRAPH-STALE-BASELINE-001 | Stale Graph Index Baseline Fallback | done | `merger/lenskit/retrieval/query_core.py`, `merger/lenskit/tests/test_graph_rerank.py`, `merger/lenskit/tests/test_graph_eval.py`, `merger/lenskit/tests/test_retrieval_query.py`, `merger/lenskit/tests/test_retrieval_eval.py`, `docs/architecture/graph-runtime-contract.md`, `docs/proofs/graph-stale-index-baseline-fallback-proof.md` | Ein erkannter `stale_or_mismatched` Graph bleibt diagnostisch sichtbar, verändert aber weder Graph-Bonus noch Entrypoint-Bonus, Test-Penalty, Reihenfolge, Scores oder Evidence-Basis. Frische opt-in Graph-Rankings bleiben unverändert. |
 | TASK-GRAPH-PROVENANCE-COHERENCE-001 | Graph Provenance-Coherent Compilation | done | `merger/lenskit/architecture/graph_source_validation.py`, `merger/lenskit/architecture/graph_index.py`, `merger/lenskit/core/merge.py`, `merger/lenskit/cli/cmd_architecture.py`, `merger/lenskit/tests/test_graph_index.py`, `merger/lenskit/tests/test_graph_bundle_integration.py`, `merger/lenskit/tests/test_graph_bundle_manifest_provenance.py`, `merger/lenskit/tests/test_cli_architecture_graph_index.py`, `docs/proofs/graph-provenance-coherent-compilation-proof.md` | Beide Compiler-Quellen werden vor der Distanzberechnung gegen Draft-07 validiert; gleiche nichtleere Run-ID und gleicher Dump-Index-Hash sind Pflicht. Der Bundlepfad bindet explizit an aktuellen Lauf und finalisierten Dump-Index und propagiert strukturierte Fehler. Automatische Source-Erzeugung, Default-Ranking-Nutzung, Graph-Vollständigkeit und Runtime-Kausalität bleiben Nicht-Ziele. |
+| TASK-GRAPH-BUNDLE-SOURCES-001 | Single-Repo Bundle-Bound Graph Sources | done | `merger/lenskit/architecture/bundle_sources.py`, `merger/lenskit/core/merge.py`, `merger/lenskit/tests/test_graph_bundle_sources.py`, `merger/lenskit/tests/test_graph_bundle_integration.py`, `docs/proofs/graph-bundle-source-production-proof.md` | Produces Graph and Entrypoints sources with current bundle provenance for unambiguous single-repo outputs; preserves fail-closed partial pairs and explicitly leaves multi-repo identity out of scope. |
